backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
paho-mqtt==1.6.1
certifi==2024.6.16